Integrate verb graph into memory
