+++ memory.py
@@ -2,7 +2,7 @@ import sqlite3
 import math
 import time
 import re
-from collections import Counter
+from collections import Counter, defaultdict
 from typing import List, Tuple, Dict
 
 DB_PATH = 'memory.db'
@@ -48,6 +48,35 @@ def metrics(words: List[str], vocab: Dict[str, int]) -> Tuple[float, float, str]
     perplexity = 2 ** entropy
     resonance_word = min(words, key=lambda w: vocab.get(w, 1)) if words else ''
     return entropy, perplexity, resonance_word
+
+
+class VerbGraph:
+    """Track verbs and trailing punctuation with edge weights."""
+
+    def __init__(self) -> None:
+        # verb -> punctuation -> count
+        self.edges: Dict[str, Counter] = defaultdict(Counter)
+
+    def add_sentence(self, text: str) -> None:
+        """Update graph counts for verbs followed by punctuation.
+
+        Parameters
+        ----------
+        text: str
+            Sentence to process.
+        """
+        tokens = re.findall(r"\b\w+\b|[.!?]", text.lower())
+        for i in range(len(tokens) - 1):
+            word, nxt = tokens[i], tokens[i + 1]
+            if nxt in ".!?":
+                self.edges[word][nxt] += 1
+
+    def preferred_punct(self, verb: str) -> str:
+        """Return punctuation most frequently following *verb*."""
+        counts = self.edges.get(verb.lower())
+        if counts:
+            return counts.most_common(1)[0][0]
+        return "."
 
 # Ensure the database exists when the module is imported
 init_db()
